+++ src/stories/Modal.stories.tsx
@@ -4,7 +4,7 @@ import Modal from '~/components/Modal';
 
 import { RequestErrorTemplate } from '~/components/Modal/templates/RequestError';
 
-import { handleModalRender } from '~/helpers/handleModal';
+import { handleModalRender } from '~/utils/handleModal';
 
 type ModalType = typeof Modal;
